Setup server-side with Express, MySQL, and Nodemon
Install the necessary node modules for the server side of the application, including Express and MySQL.

Additionally, it was added Nodemon to automatically restart the node application when file changes are detected in the directory.

Configur Nodemon to be used during development by creating an entry on package.json scripts.
- This allows developers to run the back-end node server using Nodemon and avoid restarting the server for every change in files. 

Signed-off-by: Fabio Benjovengo <fabio.benjovengo@gmail.com>