add body-parser node module
The `body-parser` module was installed to enable parsing of JSON objects sent between the front-end and back-end. When sending data from the front-end to the back-end using HTTP requests, the data is sent in the request body as a JSON object. In order to access this data on the back-end, the `body-parser` middleware must be used to parse the JSON object and make it available on the `request` object.

Without the `body-parser` middleware, the back-end would not be able to access the data sent in the request body, which would prevent the front-end and back-end from communicating effectively. By installing and configuring `body-parser`, we can now send and receive JSON data between the front-end and back-end, which is necessary for implementing features such as user authentication and database interactions.

Signed-off-by: Fabio Benjovengo <fabio.benjovengo@gmail.com>