Use the cors and the body-parser node modules in the express app
The cors and body-parser node modules were added to the express app to enable communication between the front-end and back-end servers and to parse the request body as a JSON object. Additionally, the extended option in the bodyParser middleware was set to true to enable the parsing of nested objects.

Signed-off-by: Fabio Benjovengo <fabio.benjovengo@gmail.com>