add code for sending product data to back-end via POST request
Add JavaScript code to send a POST request to the back-end server via Axios. The request sends information about a new product, including its name, description, price, and creation and update dates. The axios.post() method is used to make the request, with the product data as the request body. If the request is successful, the console will log a message confirming the insert. If an error occurs, the console will log the error message.

Signed-off-by: Fabio Benjovengo <fabio.benjovengo@gmail.com>