Add API endpoint to fetch data from database
Adds an endpoint /api/get that handles an HTTP GET request to extract information from the database. The app.get() method is used to define the endpoint, which takes two arguments: the request object req and the response object res.

The function inside the endpoint sends a SELECT statement to the database using db.query(), which selects all information from the store_products table. If the query is successful, the result is sent back to the front-end using res.send().

Signed-off-by: Fabio Benjovengo <fabio.benjovengo@gmail.com>