Add GET request to load products from the database
The useEffect hook is used to execute the function that loads the product list from the database, via a GET request to the backend server, when the component is mounted. The axios module is used to handle the HTTP request, with the response data set to update the state variable productsList. The state is initialized as an empty array.

- The useEffect hook runs when the component is mounted. It sends a GET request to the backend server to retrieve the list of products from the database.
- Use axios.get() method to send the GET request to the endpoint /api/get.
- If the request is successful, the setProductsList method is called to update the state of the products list with the response data.
- The [] empty array passed as a second argument to useEffect ensures that the hook is only executed once, when the component is mounted.

Signed-off-by: Fabio Benjovengo <fabio.benjovengo@gmail.com>