add HTTP DELETE route for deleting a product
Add a new route to handle HTTP DELETE requests for deleting a product from the database. The route takes in a productID parameter, which is used to fetch the ID of the product to be deleted from the store_products SQL table. The corresponding SQL DELETE statement is then executed to delete the product from the table.

The route is implemented in the app.delete() function and takes in the request data received from the front-end as req and sends back the response data to the front-end as res. Any errors encountered during the query execution are logged to the console.

Signed-off-by: Fabio Benjovengo <fabio.benjovengo@gmail.com>