Set update date of newly added product to null
When a new product is added to the database, it doesn't have an update date yet. This commit sets the update date of new products to null to reflect this fact. By doing this, we're ensuring that the database accurately reflects the state of the products and avoids any confusion about when a product was last updated.

Signed-off-by: Fabio Benjovengo <fabio.benjovengo@gmail.com>