add react-router-dom to the project dependencies
Add the `react-router-dom` npm module to the project dependencies using `npm install react-router-dom`. This package provides tools for routing in React applications, allowing for the creation of multiple pages and dynamic URLs.

The module will be used to manage navigation within the application and ensure that different components are rendered based on the current URL. By adding `react-router-dom`, we improve the functionality of the application and make it easier to build a more complex UI.

Signed-off-by: Fabio Benjovengo <fabio.benjovengo@gmail.com>