add boilerplate to display a single product information
Introduction of a new component, SingleProductCard, to display information about a product retrieved from a database. The component takes in a product object as a prop and renders its various properties, such as product name, description, price, creation and update dates, in a formatted manner. This allows for a consistent and modular approach to displaying product information in the app.
- The code is also accompanied by a stylesheet for styling the component.

Signed-off-by: Fabio Benjovengo <fabio.benjovengo@gmail.com>