add overlay functionality to display and edit product information
Add the functionality to display an overlay when a product card is clicked. The overlay shows the product information, and allows the user to edit and update the product information. The product information is stored in a database, and the update and delete functionality is implemented using axios.

The overlay is triggered by the displayOverlay() function which sets the overlayVisible state to true. The product information is displayed in the overlay, and the user can update the information using the input fields. The updateProduct() function updates the information in the database, and clears the input fields and closes the overlay when the user clicks the "Update" button.

The deleteProduct() function allows the user to delete a product from the database by sending a POST request to the API.

Signed-off-by: Fabio Benjovengo <fabio.benjovengo@gmail.com>