add instructions to run the project
Add details to README.md file to provide clear and concise instructions on how to run the project with an Express server on ./server folder, and a React app in ./client folder. The README.md also explains that a MySQL server is required on port 33061, and that the server is running a schema called ChallengeDB with two tables: store_products and shopping_history.

The README.md also provides instructions on how to install the necessary node modules and how to run the project as a whole.

Signed-off-by: Fabio Benjovengo <fabio.benjovengo@gmail.com>