add endpoint to get specific product by id
Add an endpoint to the RESTful API that allows the front-end to get information about a specific product by its ID. The endpoint is implemented as a GET request that extracts information from the 'store_products' SQL table using a SELECT statement.

It also includes error handling to ensure that if the requested product is not found or if there is an error fetching the data from the database, an appropriate error response is sent back to the front-end. This feature improves the functionality of the API.

Signed-off-by: Fabio Benjovengo <fabio.benjovengo@gmail.com>