add error handling to CRUD endpoints
Add error handling to the CRUD endpoints for products.
Errors are returned as JSON responses with appropriate status codes.
The error messages are also descriptive and provide helpful information
to the client.

The following changes were made:
- If there is an error fetching products from the database, return a 500 error
- If a product with the specified ID is not found, return a 404 error
- If there are missing required fields when creating or updating a product, return a 400 error
- If there is an error creating a product in the database, return a 500 error
- If there is an error updating a product in the database, return a 500 error
- If a product with the specified ID cannot be found when updating or deleting, return a 404 error
- If there is an error deleting a product from the database, return a 500 error

These changes ensure that the API provides useful and meaningful error messages to the client, making it easier for them to debug and troubleshoot issues.

Signed-off-by: Fabio Benjovengo <fabio.benjovengo@gmail.com>