add design decisions to README file
Add a description of the design decisions made for the project, including the use of body-parser, cors, express, mysql, and nodemon. Additionally, it includes a description of the decision to use the axios, bootstrap, and reactstrap modules, which were chosen to simplify communication with the server-side API and provide a user-friendly and responsive UI experience.

Signed-off-by: Fabio Benjovengo <fabio.benjovengo@gmail.com>