add comments to demonstrate POST and GET requests using Axios
Add comments to the code that demonstrates how to send a POST and a GET request to a server using the Axios library in JavaScript. The comments include information about the endpoints used for the requests, the form data passed in the POST request, and how to retrieve data from the server using the GET request.

- Change the variable names for the endpoints for the GET and POST requests to be more descriptive of their respective functions in the code.

Signed-off-by: Fabio Benjovengo <fabio.benjovengo@gmail.com>